Forced Unmount of a filesystem
The patch was originally written by Dave Jiang in Montavista.
This is the update of the Dave's patch. I cleaned up the unmapping
vmas and done some bug fixes.
If fumount closes the file, while a process has a reference to
a file, the file is freed suddenly. To solve this problem,
a new count variable is added in the struct file. This count is
increased when a system call starts and decreased when system
call exits. The related system calls are file operations like
read, write, seek, etc. This count makes fumount know that the
file operation has a reference to a file. Fumount does not close
the file as long as the file operation is in progress.

Signed-off-by: Naresh Bhat <nareshgbhat@gmail.com>